Add configurable creating of .json files

+++ models/json/thread.js
@@ -1,5 +1,6 @@
 const FS = require('../../helpers/fs'),
-  db = require('../db/thread');
+    db = require('../db/thread'),
+    config = require('../../helpers/config');
 
 let thread = module.exports = {};
 
@@ -11,8 +12,9 @@ let thread = module.exports = {};
 thread.create = async function(fields) {
   let query = await db.create(fields);
   let out = { board: fields['boardName'], thread: query['insertId'] };
-  query = await db.read(out.board, out.thread);
-  FS.writeSync(`${out.board}/res/${out.thread}.json`, JSON.stringify(query));
+  thread = await db.read(out.board, out.thread);
+  if (config('fs.cache.json'))
+    FS.writeSync(`${out.board}/res/${out.thread}.json`, JSON.stringify(thread));
   return out;
 };
 
@@ -23,12 +25,20 @@ thread.create = async function(fields) {
  * @return {Object} query
  */
 thread.read = async function(board, thread_id) {
-  let pattern = `${board}/res/${thread_id}.json`, query;
-  if (FS.existsSync(pattern))
-    query = JSON.parse(FS.readSync(pattern)); //TODO: Read file async
-  else
-    query = await db.read(board, thread_id);
-  return query;
+  let pattern = `${board}/res/${thread_id}.json`, thread;
+  if (config('fs.cache.json')) {
+    if (FS.existsSync(pattern)) {
+      let file = FS.readSync(pattern); //TODO: Read file async
+      try {
+        thread = JSON.parse(file);
+      } catch (e) {
+        //
+      }
+    }
+  }
+  if (!thread)
+    thread = await db.read(board, thread_id);
+  return thread;
 };
 
 /**
@@ -39,16 +49,15 @@ thread.read = async function(board, thread_id) {
  * @return {Object} query
  */
 thread.update = async function(board, thread_id, post_id) {
-  let query = await db.update(board, thread_id, post_id),
-      file = FS.readSync(`${board}/res/${thread_id}.json`);
-  if (query.length < 1)
-    return false;
+  let query = await db.update(board, thread_id, post_id);
+  if (query.length < 1 || !config('fs.cache.json'))
+    return query;
+  let file = FS.readSync(`${board}/res/${thread_id}.json`);
   //TODO: check for existing posts and replace them
   // for now use thread.regenerateJSON()
   file = JSON.parse(file);
   Array.prototype.push.apply(file, query);
   FS.writeSync(`${board}/res/${thread_id}.json`, JSON.stringify(file));
-  return query;
 };
 
 /**
@@ -58,8 +67,14 @@ thread.update = async function(board, thread_id, post_id) {
  * @return {Object} query
  */
 thread.regenerateJSON = async function(board, thread_id) {
+  let pattern = 'fs.cache.json';
+  if (!config(pattern)) {
+    console.log(`Ni-paa~! Enable '${pattern}' in your config to use this feature.`);
+    return false;
+  }
   let query = await db.regenerateJSON(board, thread_id);
   FS.writeSync(`${board}/res/${thread_id}.json`, JSON.stringify(query));
+  return query;
 };
 
 /**
@@ -69,7 +84,8 @@ thread.regenerateJSON = async function(board, thread_id) {
  * @return {Boolean}
  */
 thread.delete = async function(board, thread_id) {
-  let query = await db.delete(board, thread_id);
-  FS.unlinkSync(`${board}/res/${thread_id}.json`);
+  let query = await db.delete(board, thread_id); //TODO: Check password
+  if (config('fs.cache.json'))
+    FS.unlinkSync(`${board}/res/${thread_id}.json`);
   return query;
 };
